Final Submission
Refactored the cases in the switch statement into a function since they
all did the same thing.  Just needed to pass in an argument for what
the class name should be changed to.

Also used .attr() instead of addClass() to change the class name for
each background.

Took out the log statements as well.